requirements.txt
flask>=3.0.0
reportlab>=4.1.0
pillow>=10.0.0
werkzeug>=3.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
